# onyx-procurement/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ONYX Procurement API
  description: |
    API for the ONYX Procurement System — covers procurement (suppliers,
    RFQ, quotes, POs, subcontractors), VAT (PCN836), annual tax forms,
    bank reconciliation, payroll/wage slips, and WhatsApp webhooks.
  version: 1.0.0
  contact:
    name: Techno-Kol Ops
servers:
  - url: http://localhost:3100
    description: Local development
  - url: https://api.onyx.example.com
    description: Production placeholder
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    VatProfile:
      type: object
      additionalProperties: true
    VatPeriod:
      type: object
      additionalProperties: true
      properties:
        id: { type: integer }
        period_start: { type: string, format: date }
        period_end: { type: string, format: date }
        period_label: { type: string }
        status: { type: string, enum: [open, closing, submitted] }
    TaxInvoice:
      type: object
      additionalProperties: true
      properties:
        id: { type: integer }
        direction: { type: string, enum: [input, output] }
        invoice_number: { type: string }
        invoice_date: { type: string, format: date }
        net_amount: { type: number }
        vat_amount: { type: number }
        gross_amount: { type: number }
        is_asset: { type: boolean }
        is_zero_rate: { type: boolean }
        is_exempt: { type: boolean }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State conflict (already decided/closed/etc.)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /api/vat/profile:
    get:
      tags: [VAT]
      summary: Get the company tax profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile: { $ref: "#/components/schemas/VatProfile" }
    put:
      tags: [VAT]
      summary: Upsert the tax profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VatProfile" }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile: { $ref: "#/components/schemas/VatProfile" }
  /api/vat/periods:
    get:
      tags: [VAT]
      summary: List VAT periods
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 24 }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  periods:
                    type: array
                    items: { $ref: "#/components/schemas/VatPeriod" }
    post:
      tags: [VAT]
      summary: Open a new period
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VatPeriod" }
      responses:
        "201":
          description: Period created
          content:
            application/json:
              schema:
                type: object
                properties:
                  period: { $ref: "#/components/schemas/VatPeriod" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/vat/periods/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [VAT]
      summary: Period detail with computed totals
      responses:
        "200":
          description: Period + live-computed totals
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404": { $ref: "#/components/responses/NotFound" }
  /api/vat/periods/{id}/close:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      tags: [VAT]
      summary: Compute and lock a period
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/vat/periods/{id}/submit:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      tags: [VAT]
      summary: Build PCN836 file and record the submission
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                submission_type: { type: string, default: initial }
                submission_method: { type: string, default: shamat }
                submitted_by: { type: string }
      responses:
        "201":
          description: Submission recorded + file archived
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412":
          description: Company tax profile not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: PCN836 validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/vat/periods/{id}/pcn836:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [VAT]
      summary: Download PCN836 file (windows-1255 binary)
      responses:
        "200":
          description: File stream
          content:
            text/plain:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
        "410":
          description: File referenced but missing from archive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/vat/invoices:
    get:
      tags: [VAT]
      summary: List tax invoices
      parameters:
        - in: query
          name: direction
          schema: { type: string, enum: [input, output] }
        - in: query
          name: period_id
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items: { $ref: "#/components/schemas/TaxInvoice" }
    post:
      tags: [VAT]
      summary: Record a tax invoice (auto-computes VAT if missing)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaxInvoice" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoice: { $ref: "#/components/schemas/TaxInvoice" }
...